+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## WIP
+- Add ability to pause from row
+- Javascript improvements
 
 ## 1.0.0
 - Initial release